Add index.php- (home page) and shop page
in this step, I uploaded the home and shop pages to the workspace(only the design part of both pages). All the pages are coded by me in the previous week but I forget to upload them to the git hub. the slider images and the product images are temporary. product images and slider images will be changed in the future steps.